requirements.txt
aiogram>=3.15.0
python-dotenv==1.0.0
sqlalchemy>=2.0.36
aiohttp==3.9.5
openai>=1.109.0
